Minor changes to Score/Hud-T prefab.
The way the prefab was working didn't play well with how the score works at this point in time, so I temporarily changed it back so we have something functional for this upcoming sprint. We absolutely need to change this at some point in the future to get an overall system that works.